Rudimentary configuration
I tend to fail a lot so let's have something to fall back on.

I had a huge performance issue among other annoying thing.

It was doom-modeline's fault. I suspect it's link with my alacritty, discord,
anydesk problem. it seems to be a memory leak? or a problem with the GPU as
discord complained about. I'm not sure what happened. I must have updated my
system and something unexpected changed. I had this problem but it "solved" by
itself with alacritty but it happened again after an update.